This is a feet to inches conversion project.

+++ src/HelloNetbeans.java
@@ -12,8 +12,8 @@ public class HelloNetbeans
         Scanner s1 = new Scanner(System.in);
         System.out.println("Welcome!");
         System.out.println("You will now enter an integer to convert from feet to inches:");
-        int a;
-        a = s1.nextInt();
+        double a;
+        a = s1.nextDouble();
         System.out.println("That is " + a + " feet");
         System.out.println("The converted integer is " + a*12 + " inches");
     }
